@@ -30,9 +30,11 @@
     "@babel/preset-env": "^7.16.0",
     "@babel/preset-react": "^7.16.0",
     "@babel/preset-typescript": "^7.17.12",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.2",
+    "@testing-library/jest-dom": "^5.16.0",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/react-hooks": "^8.0.0",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
     "@typescript-eslint/parser": "^5.0.0",
     "babel": "^6.23.0",
@@ -76,9 +78,9 @@
     "webpack-manifest-plugin": "^2.2.0"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.8.3",
-    "@emotion/cache": "^11.4.0",
-    "@emotion/react": "^11.4.1",
+    "@chakra-ui/react": "^2.2.0",
+    "@emotion/cache": "^11.9.3",
+    "@emotion/react": "^11.9.3",
     "@emotion/styled": "^11",
     "axios": "^0.26.0",
     "bootstrap-3-typeahead": "^4.0.2",
@@ -91,14 +93,14 @@
     "datatables.net": "^1.11.4",
     "datatables.net-bs": "^1.11.4",
     "eonasdan-bootstrap-datetimepicker": "^4.17.47",
-    "framer-motion": "^4",
+    "framer-motion": "4.1.17",
     "jquery": ">=3.5.0",
     "jshint": "^2.13.4",
     "lodash": "^4.17.21",
     "moment-timezone": "^0.5.34",
     "nvd3": "^1.8.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
     "react-query": "^3.34.16",
     "react-router-dom": "^6.3.0",
